requirements/local.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
-r base.txt
alabaster==0.7.12
argon2-cffi==21.1.0
asgiref==3.4.1; python_version >= '3.6'
astroid==2.8.0; python_version ~= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
babel==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
backcall==0.2.0
backports.entry-points-selectable==1.1.0; python_version >= '2.7'
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1; python_full_version >= '3.6.1'
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.1; python_version >= '3.6'
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
decorator==5.1.0; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.3.3
django-allauth==0.45.0
django-cors-headers==3.8.0
django-coverage-plugin==2.0.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-extensions==3.1.3
django-model-utils==4.1.1
django-redis==5.0.0
django-stubs-ext==0.3.1; python_version >= '3.6'
django-stubs==1.8.0
django==3.1.13
djangorestframework==3.12.4
docutils==0.17.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
drf-yasg==1.20.0
factory-boy==3.2.0
faker==8.13.2; python_version >= '3.6'
filelock==3.0.12
flake8-isort==4.0.0
flake8==3.9.2
hiredis==2.0.0
identify==2.2.15; python_full_version >= '3.6.1'
idna==3.2; python_version >= '3'
imagesize==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.27.0; python_version >= '3.7'
isort==5.9.3; python_version < '4.0' and python_full_version >= '3.6.1'
itypes==1.2.0
jedi==0.18.0; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
livereload==2.6.3
markupsafe==2.0.1; python_version >= '3.6'
matplotlib-inline==0.1.3; python_version >= '3.5'
mccabe==0.6.1
mypy-extensions==0.4.3
mypy==0.910
nodeenv==1.6.0
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
parso==0.8.2; python_version >= '3.6'
pathspec==0.9.0
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==8.3.2
platformdirs==2.3.0; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
pre-commit==2.15.0
prompt-toolkit==3.0.20; python_full_version >= '3.6.2'
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycodestyle==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyflakes==2.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.10.0; python_version >= '3.5'
pyjwt[crypto]==2.1.0; python_version >= '3.6'
pylint-django==2.4.4
pylint-plugin-utils==0.6
pylint==2.11.1; python_version ~= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-django==4.4.0
pytest-sugar==0.9.4
pytest==6.2.5
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
redis==3.5.3
regex==2021.8.28
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.16; python_version >= '3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
snowballstemmer==2.1.0
sphinx-autobuild==2021.3.14
sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2; python_version >= '3.5'
sphinxcontrib-devhelp==1.0.2; python_version >= '3.5'
sphinxcontrib-htmlhelp==2.0.0; python_version >= '3.6'
sphinxcontrib-jsmath==1.0.1; python_version >= '3.5'
sphinxcontrib-qthelp==1.0.3; python_version >= '3.5'
sphinxcontrib-serializinghtml==1.1.5; python_version >= '3.5'
sqlparse==0.4.2; python_version >= '3.5'
termcolor==1.1.0
testfixtures==6.18.2
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==1.2.1; python_version >= '3.6'
tornado==6.1; python_version >= '3.5'
traitlets==5.1.0; python_version >= '3.7'
typing-extensions==3.10.0.2; python_version < '3.10'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
virtualenv==20.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wcwidth==0.2.5
werkzeug==1.0.1
wrapt==1.12.1
